requirements.txt
altgraph==0.10.2
Babel==2.1.1
bdist-mpkg==0.5.0
blinker==1.4
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7
django-toolbelt==0.0.1
djangorestframework==3.3.1
extensions==0.3
Flask==0.10.1
Flask-Babel==0.9
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.1
Flask-WhooshAlchemy==0.55a0
Flask-WTF==0.12
flup==1.0.2
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
macholib==1.5.1
MarkupSafe==0.23
modulegraph==0.10.4
numpy==1.8.0
Pillow==3.0.0
psycopg2==2.6.1
py2app==0.7.3
pymongo==3.1.1
pyOpenSSL==0.13.1
pyparsing==2.0.1
python-dateutil==1.5
python-openid==2.2.5
pytz==2013.7
requests==2.8.1
six==1.4.1
speaklater==1.3
SQLAlchemy==1.0.9
static3==0.6.1
Twisted==13.2.0
Unipath==1.1
Werkzeug==0.11.2
Whoosh==2.7.0
WTForms==2.0.2
xattr==0.6.4
zope.interface==4.1.1